+++ Source/AbletonLink.h
@@ -57,7 +57,7 @@ public:
    void SetEnabled(bool enabled) override { mEnabled = enabled; }
 
    void CheckboxUpdated(Checkbox* checkbox, double time) override;
-   void FloatSliderUpdated(FloatSlider* slider, float oldVal, double time) override;
+   void FloatSliderUpdated(FloatSlider* slider, double oldVal, double time) override;
    void ButtonClicked(ClickButton* button, double time) override;
    bool IsEnabled() const override { return mEnabled; }
 
@@ -73,7 +73,7 @@ private:
    float mWidth{ 200 };
    float mHeight{ 20 };
 
-   float mOffsetMs{ 0 };
+   double mOffsetMs{ 0 };
    FloatSlider* mOffsetMsSlider{ nullptr };
    ClickButton* mResetButton{ nullptr };
