+++ Source/BoundsToPulse.h
@@ -50,7 +50,7 @@ public:
 
    void OnTransportAdvanced(float amount) override;
 
-   void FloatSliderUpdated(FloatSlider* slider, float oldVal, double time) override;
+   void FloatSliderUpdated(FloatSlider* slider, double oldVal, double time) override;
 
    float GetValue() const { return mValue; }
    FloatSlider* GetSlider() { return mSlider; }
@@ -64,7 +64,7 @@ private:
    }
 
    FloatSlider* mSlider;
-   float mValue;
+   double mValue;
 
    PatchCableSource* mMinCable{};
    PatchCableSource* mMaxCable{};
