+++ Source/ClipArranger.h
@@ -51,7 +51,7 @@ public:
 
    void FilesDropped(std::vector<std::string> files, int x, int y) override;
 
-   void FloatSliderUpdated(FloatSlider* slider, float oldVal, double time) override;
+   void FloatSliderUpdated(FloatSlider* slider, double oldVal, double time) override;
    void ButtonClicked(ClickButton* button, double time) override;
    void CheckboxUpdated(Checkbox* checkbox, double time) override;
